Add additional shader that always casts original shadows
Added an additional shader called DitherTransparentOriginalShadow. Instead of a dithered shadow, the original shadow will always be cast regardless of the opacity. An additional node was added to the shader graph, using the technique referenced by Unity on this forum post, that sets IsShadowCaster to true: https://forum.unity.com/threads/alpha-clip-dither-and-keep-original-shadows.781412/#post-5212907